ops-distiller-forge/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-distiller-forge"
version = "0.1.0"
description = "② 炼金工坊：DSPy 蒸馏、Pydantic 母版、手册与 Graphiti Episode 投影、Metric/GEPA、Agent Manifest"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dspy = [
  "dspy-ai>=2.5.0",
  "openai>=1.40.0",
]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
ops-distiller = "ops_distiller_forge.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ops_distiller_forge = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
